Revert "crawler.ts marge"
