+++ src/@types/monitorItem.ts
@@ -1,4 +1,6 @@
-export type MonitorItem = {
+type MonitorItem = {
   id: string | number | Buffer;
   value: string;
 };
+
+export default MonitorItem
